requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
numpy==1.26.4
scikit-learn==1.5.2
tensorflow-cpu==2.17.0
